@@ -1,23 +1,31 @@
 {
   "name": "project",
+  "homepage": "https://gitname.github.io/GIS-Projekt",
   "dependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "npm-check-updates": "^13.0.1",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
   },
   "devDependencies": {
-    "babel-core": "^6.0",
-    "babel-loader": "^7.0",
-    "babel-preset-env": "*",
-    "babel-preset-react": "*",
+    "@babel/core": "^7.18.0",
+    "@babel/plugin-proposal-class-properties": "^7.17.12",
+    "@babel/plugin-transform-runtime": "^7.18.0",
+    "@babel/preset-env": "^7.18.0",
+    "@babel/preset-react": "^7.17.12",
+    "babel-loader": "^8.2.5",
     "css-loader": "^6.7.1",
+    "gh-pages": "^4.0.0",
+    "html-webpack-plugin": "^5.5.0",
     "style-loader": "^3.3.1",
     "webpack": "^5.72.1",
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.9.0"
   },
   "scripts": {
-    "build": "webpack",
+    "build:dev": "webpack --mode development",
+    "build:prod": "webpack --mode production",
     "start": "webpack-dev-server --mode development --open --hot",
+    "deploy": "npm run build:prod && gh-pages -d dist",
     "watch": "webpack -w"
   }
 }
